refactor: scope seed
