+++ src/libs/auth/src/delegation/get.rs
@@ -1,4 +1,4 @@
-use crate::seed::calculate_seed;
+use crate::delegation::seed::calculate_seed;
 use crate::services::read_state;
 use crate::strategies::AuthCertificateStrategy;
 use crate::types::interface::{
