+++ src/libs/auth/src/delegation/mod.rs
@@ -1,5 +1,6 @@
 mod get;
 mod prepare;
+mod seed;
 
 pub use get::*;
 pub use prepare::*;
