+++ src/libs/auth/src/delegation/prepare.rs
@@ -1,5 +1,5 @@
 use crate::constants::DEFAULT_EXPIRATION_PERIOD_NS;
-use crate::seed::calculate_seed;
+use crate::delegation::seed::calculate_seed;
 use crate::services::mutate_state;
 use crate::strategies::AuthCertificateStrategy;
 use crate::types::interface::{
