+++ src/libs/auth/src/lib.rs
@@ -2,7 +2,6 @@ mod constants;
 pub mod delegation;
 mod memory;
 pub mod runtime;
-pub mod seed;
 mod services;
 pub mod strategies;
 pub mod types;
